libublk: exe: support async uring op
The basic idea is to pass current cqe into one "stack variable"
which is simulated by one small Rust thread_local!() variable.

Then future .poll() of uring OP can find this incoming cqe, and compare
cqe.user_data() with the pre-saved userdata(key) in future, and check
if it is completed.

This way is pretty flexible, and very fast, and easy to extend to
support multi-shot CQE.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>