libublk: ctrl: add UblkDev::kill_dev()
UblkDev::del() can cause deadlock easily, that is why run_target()
calls device_fn() in another thread context. The device fn handler
is joined after this control device is done, so the thread context
for running device_fn() may take long time.

This way wastes resource, and can't work well in case of tens of
thousands of devices.

Add UblkDev::kill_dev() which can be called safely in device_fn()
passed to run_target(), and prepare for not running device_fn()
in another context.

One thing about kill_dev() is that the device may not be removed
really, and the device can be in-use, and it is good to call it
to remove device inside device_fn() of run_target().

For other cases, Ctrl::del_dev() is just fine.

Also mark start()/stop()/del() as private.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>